+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Always satisfy the collateral contract input in the loan repayment transaction correctly.
   We used to assume that it was located at index `1`, but this could change if the number of inputs used to cover the repayment output was greater than 1.
 
+### Changed
+
+- Lender sets collateral contract timelock, as opposed to the borrower.
+- Deprecated all APIs which use `LoanRequest`, to give users more control over how they want to serialize the parameters inside loan request.
+  We've also added a new API `build_loan_transaction` on `Lender0`, which wraps the now deprecated `interpret`.
+
 ## [0.2.0] - 2021-07-30
 
 ### Added